Testing Token

+++ LKS 2025/restapi/routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\AuthController;
+use App\Http\Controllers\PlayerController;
 use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -14,6 +15,14 @@ Route::post('/users/auth', [AuthController::class, 'login']);
 Route::post('/users', [UserController::class, 'createUser']);
 Route::get('/users', [UserController::class, 'getAllUser']);
 Route::get('/users/{id}', [UserController::class, 'getUser']);
-Route::put('/users/{id}/update', [UserController::class, 'updateUser']);
-Route::delete('/users/{id}/delete', [UserController::class, 'deleteUser']);
+Route::put('/users/{id}/update', [UserController::class, 'updateUser'])->middleware('auth:sanctum');
+Route::delete('/users/{id}/delete', [UserController::class, 'deleteUser'])->middleware('auth:sanctum');
+Route::get('/player', [PlayerController::class, 'getAllPlayers']);
+Route::post('/player/post', [PlayerController::class, 'createNewPlayer'])->middleware('auth:sanctum');
+Route::get('/test-token', function (Request $request) {
+    return response()->json([
+        'user' => $request->user(),
+        'token_valid' => $request->user() ? true : false
+    ]);
+})->middleware('auth:sanctum');
